+++ WoWsShipBuilder.Core/Settings/DispersionPlotSettings.cs
@@ -6,6 +6,8 @@
 
         public double AimingRange { get; set; } = 10.0;
 
+        public double ShootingRange { get; set; } = 10.0; // this is for shell trajectory chart
+
         public bool IsVertical { get; set; } = true;
     }
 }